requirements.txt
flask
flask-mysql
pyopenssl
